Refactor vocabulary retrieval from Anki + configuration
First of all, the previously hardcoded values of Anki DB location,
notetypes that contain vocabulary and export dialog default folder are
all made configurable through a config file.

If the configuration file is not present in the DATA_DIR, the user is
prompted to fill in the values at startup.

Instead of configuring individual fields on the notetypes that contain
vocabulary, now all the text contained on a note is considered source of
vocabulary. A note's text is just all fields appended, words are
automatically extracted through segmenting text and filtering for valid
words that appear in the dictionary simplified<->traditional mappings.
All vocabulary is in simplified character form.